webui: do not submit form on enter
THe QR and barcode reader will finish every input with a return/enter key.
Do not submit the form when that happens as more data needs to be put in.

Reloading the form when scanning in the ID field is still intended. It brings
up the stored data for that ID.

Signed-off-by: Stefan Nuernberger <kabelfrickler@gmail.com>